+++ scripts/state_machine.pl
@@ -1,7 +1,7 @@
 #!/perl -i.bak
 ################################################################################
 #
-# $Id$
+#
 #
 # Copyright 2011 Crown copyright (c)
 # Land Information New Zealand and the New Zealand Government.