+++ src/bde_copy_gzip.cpp
@@ -1,23 +1,25 @@
 /***************************************************************************
- $Id$
+
 
  Copyright 2011 Crown copyright (c)
  Land Information New Zealand and the New Zealand Government.
  All rights reserved
 
- This program is released under the terms of the new BSD license. See 
+ This program is released under the terms of the new BSD license. See
  the LICENSE file for more information.
 ****************************************************************************/
 
 #include "bde_copy_gzip.h"
 #include "bde_copy_funcs.h"
 #include "zlib.h"
+#include <string.h>
+#include <errno.h>
 
 gzipbuff::gzipbuff( char *name, int zipbuffsize ) : readbuff( name, filebuffsize )
 {
     gzFile p = gzopen(name,"rb");
     gz = (void *)p;
-    if( gz ) 
+    if( gz )
     {
         setstatus(OK);
 #if ZLIB_VERNUM >= 0x1235
@@ -25,7 +27,7 @@ gzipbuff::gzipbuff( char *name, int zipbuffsize ) : readbuff( name, filebuffsize
         gzbuffer(p,zipbuffsize);
 #endif
     }
-    else 
+    else
     {
         setstatus(OPEN_ERROR);
     }
@@ -46,9 +48,17 @@ int gzipbuff::fill()
 
 data_writer *gzip_data_writer::open( char *fname, bool append, int zipbuffsize )
 {
-    if( append && file_exists(fname)) return 0;
+    if( append && file_exists(fname)) {
+      // TODO: print detail about this failure
+      return 0;
+    }
     void *gz = gzopen(fname,"wb" );
-    if( ! gz ) return 0;
+    if( ! gz ) {
+      // TODO: find a better way to signal errors to upper levels (throw?)
+      fprintf(stderr,"Could not open %s for writing: %s\n",
+        fname, strerror(errno));
+      return 0;
+    }
 #if ZLIB_VERNUM >= 0x1235
     if( zipbuffsize == 0 ) zipbuffsize = readbuff::filebuffsize;
     gzbuffer(static_cast<gzFile>(gz),zipbuffsize);