+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes for the LINZ BDE Copy are documented in this file.
 
+## [1.3.0] - 2018-02-05
+### Changed
+- Print startup errors to stderr (#33)
+- Make base configuration file optional
+### Added
+- Add ability to output to stdout (#31)
+- Version flag (-V)
+- Support for compiling on MacOSX
+### Fixed
+- Win32 build
+- File size error when reading more than one file
+### Improved
+- Documentation improvements
+
 ## [1.2.1] - 2016-05-01
 ### Changed
 - Minor documentation changes
@@ -9,7 +23,7 @@ All notable changes for the LINZ BDE Copy are documented in this file.
 ## [1.2.0] - 2015-04-09
 ### Added
 - Added static casts to allow compilation of gzip code
-- Added script to build bde_copy.cfg.lolutf8
+- Added script to build `bde_copy.cfg.lolutf8`
 - Added additional Landonline UTF8 chars
 
 ### Changed
@@ -31,4 +45,3 @@ All notable changes for the LINZ BDE Copy are documented in this file.
 
 ### Fixed
 - Fixing remaining compiler warnings and introduced bug
-